requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1
django-phonenumber-field==6.3.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
inflection==0.5.1
jsonschema==4.9.1
phonenumbers==8.12.53
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
